% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_drought_events.R
\name{add_drought_events}
\alias{add_drought_events}
\title{Add drought period (bands or markers) to a spei time series plot}
\usage{
add_drought_events(
  p,
  drought_assessment,
  which_events = c("all", "top"),
  metric = c("duration", "severity", "intensity", "lowest_index"),
  top_n = 5,
  type = c("line", "polygon", "both"),
  line_col = "black",
  line_type = "dashed",
  pol_fill = "gray70",
  pol_alpha = 0.5,
  show_severity = FALSE
)
}
\arguments{
\item{p}{A \code{ggplot} object created by \code{plot_drought_ts()}.}

\item{drought_assessment}{A data frame with columns \code{minyear}, \code{month_peak}, \code{d_duration},
\code{d_severity}, and others, as returned by \code{droughts()}.}

\item{which_events}{Display \code{"all"} droughts or only the \code{"top"} events.}

\item{metric}{If \code{which_events = "top"}, use this column to rank events. Options:
\code{"duration"}, \code{"severity"}, \code{"intensity"}, \code{"lowest_index"}.}

\item{top_n}{Number of top events to display when \code{which_events = "top"}. Default to 5.}

\item{type}{Type of marker to display: \code{"polygon"} (shaded bands), \code{"line"} (vertical lines at peak), or \code{"both"}.}

\item{line_col}{Color of vertical line (if used).}

\item{line_type}{Line type of vertical line (e.g. \code{"dashed"}, \code{"solid"}).}

\item{pol_fill}{Fill color of polygon (if used).}

\item{pol_alpha}{Transparency of polygon fill.}

\item{show_severity}{Logical. If \code{TRUE}, plots a point and label for severity at top of each event.}
}
\value{
A modified \code{ggplot} object.
}
\description{
Adds shaded polygons, vertical lines, or both to a \code{ggplot} object to highlight drought events,
based on a drought assessment summary (as returned by \code{droughts()}). Optionally,
labels or points can be added to indicate drought severity.
}
\examples{
data(spei_granada)

result <- droughts(spei_granada, vname = "spei12", threshold = -1.28)

p <- plot_drought_ts(spei_granada, vname = "spei12")

add_drought_events(
  p,
  drought_assessment = result$drought_assessment,
  which_events = "top",
  metric = "severity",
  top_n = 3,
  type = "both",
  show_severity = TRUE
)

}
